Create Password_Crack.py
A password cracking tool in Python can be implemented using various methods, including brute-force attacks. Below is an example of a simple brute-force password cracker that tries all possible combinations of characters up to a specified length.

This example uses the SHA-256 hashing algorithm to hash a target password and then attempts to crack it by generating all possible combinations of lowercase letters and digits up to a specified maximum length. The time taken to crack the password is also recorded and printed.